MAILBOX-359 Factorize tests for UpdatedFlags ser-deserialization
